:recycle: refactor : rename RunningServiceUtils to RunningKeyUtil and rename packages to singular
