+++ src/main/java/com/run_us/server/domains/running/controller/RunningController.java
@@ -2,7 +2,7 @@ package com.run_us.server.domains.running.controller;
 
 import com.run_us.server.domains.running.controller.model.request.RunningCreateRequest;
 import com.run_us.server.domains.running.service.RunningPreparationService;
-import com.run_us.server.domains.running.service.model.JoinedParticipantsDto;
+import com.run_us.server.domains.running.service.model.JoinedParticipant;
 import com.run_us.server.global.common.SuccessResponse;
 import com.run_us.server.global.exceptions.enums.ExampleErrorCode;
 import java.util.List;
@@ -29,7 +29,7 @@ public class RunningController {
 
   @GetMapping("/{runningId}/participants")
   public SuccessResponse joinedParticipants(@PathVariable String runningId) {
-    List<JoinedParticipantsDto> joinedParticipants = runningPreparationService.getJoinedParticipants(
+    List<JoinedParticipant> joinedParticipants = runningPreparationService.getJoinedParticipants(
         runningId);
     return SuccessResponse.of(ExampleErrorCode.SUCCESS, joinedParticipants);
   }
